Properly display positive conversion if rate has positive premium
